fileStream class extract
